requirements.txt
altair @ file:///home/conda/feedstock_root/build_artifacts/altair-split_1734244716962/work
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1733247046149/work
anyio @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_anyio_1754315087/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1731096409132/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1749229904370/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1750147854206/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1754231422783/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560630059/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1754767332901/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1747811314515/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1734729835256/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
gitdb @ file:///home/conda/feedstock_root/build_artifacts/gitdb_1735887193964/work
GitPython @ file:///home/conda/feedstock_root/build_artifacts/gitpython_1753444155689/work
h11 @ file:///home/conda/feedstock_root/build_artifacts/h11_1745526374115/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_httpcore_1745602916/work
httpx @ file:///home/conda/feedstock_root/build_artifacts/httpx_1733663348460/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
jiter @ file:///Users/runner/miniforge3/conda-bld/jiter_1747609813066/work
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jsonschema_1755595646/work
jsonschema-specifications @ file:///tmp/tmpuvkyqc9y/src
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1733219683650/work
narwhals @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_narwhals_1755254908/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1755864490103/work/dist/numpy-2.3.2-cp311-cp311-macosx_10_13_x86_64.whl#sha256=fa488d16fc19a3f15f34df83a8ea5f9fe6c9da7c530b334e1b49c1b000c8865b
openai @ file:///home/conda/feedstock_root/build_artifacts/openai_1755849915357/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas @ file:///Users/runner/miniforge3/conda-bld/pandas_1755779011466/work
pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1751482042455/work
protobuf @ file:///Users/runner/miniforge3/conda-bld/protobuf_1751668396245/work/bazel-bin/python/dist/protobuf-6.31.1-cp311-abi3-macosx_10_13_x86_64.whl#sha256=afac6f41e8e7e61987964f3cc9c8a0cc086af3816ba412b8d8bafe1740fddbf4
pyarrow==21.0.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic @ file:///home/conda/feedstock_root/build_artifacts/pydantic_1749927112905/work
pydantic_core @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_pydantic-core_1746625302/work
pydeck @ file:///home/conda/feedstock_root/build_artifacts/pydeck_1738346561094/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dotenv_1750789290/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1737454694178/work
referencing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_referencing_1737836872/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1755614211359/work
rpds-py @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_rpds-py_1754570012/work
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smmap @ file:///home/conda/feedstock_root/build_artifacts/smmap_1739781697784/work
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1733244044561/work
streamlit @ file:///home/conda/feedstock_root/build_artifacts/streamlit_1755128470103/work
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1743640740204/work
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1734091811753/work
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1754732036208/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
typing-inspection @ file:///home/conda/feedstock_root/build_artifacts/typing-inspection_1747870647094/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1751643513/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
watchdog @ file:///Users/runner/miniforge3/conda-bld/watchdog_1730492939283/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
zstandard==0.23.0
